+++ coxeter/__init__.py
@@ -1,19 +1,15 @@
-from . import shapes
-from . import damasceno
 from . import symmetry
 from . import ft
 from . import shape_classes
-from . import common_shapes
+from . import shape_families
 from .shape_getters import from_gsd_type_shapes
 
 __all__ = [
-    'shapes',
-    'damasceno',
     'symmetry',
     'ft',
     'shape_classes',
-    'common_shapes',
-    'from_gsd_type_shapes'
+    'from_gsd_type_shapes',
+    'shape_families'
 ]
 
 __version__ = '0.2.0'