+++ coxeter/shape_families/__init__.py
@@ -1,0 +1,11 @@
+from .plane_shape_families import (Family323Plus, Family423, Family523,
+                                   TruncatedTetrahedronFamily)
+from .tabulated_shape_family import (TabulatedShapeFamily,
+                                     TabulatedGSDShapeFamily)
+from .common_families import RegularNGonFamily, PlatonicFamily
+from .doi_data_repositories import family_from_doi
+
+__all__ = ['TabulatedShapeFamily', 'TabulatedGSDShapeFamily',
+           'shape_repositories', 'Family323Plus', 'Family423', 'Family523',
+           'TruncatedTetrahedronFamily', 'family_from_doi',
+           'RegularNGonFamily', 'PlatonicFamily']